pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "tonydbc"
version = "1.3.0"
description = "A high-level database connector for MariaDB"
readme = "README.md"
authors = [{ name = "Michael Currie", email = "mcurrie@gmail.com" }]
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["mariadb", "database", "connector", "ODBC", "pandas"]
dependencies = [
    "iteration_utilities",
    "types-python-dateutil",
    "python-dotenv",
    "more-itertools",
    "jsonschema",
    "requests",
    "numpy>=1.23.5",
    "pandas",
    "paho-mqtt",
    "mariadb",
    "pyperclip",
    "types-pyperclip",
    "tzlocal",
    "paho-mqtt>=1.6.1",
    "filelock",
    "uuid6",
    "types-docker"
]
requires-python = ">=3.10,<4.0"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/Fling-Asia/tonydbc"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]  # Enable error, pyflakes, and warning categories, organize imports
extend-ignore = [
  "E203",  # whitespace before ':' (Black handles this?  But we don't use black anymore; oh well)
  "E501",  # line too long (let formatter handle this)
]

[tool.ruff.format]
quote-style = "double"  # Or "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.bumpver]
current_version = "1.3.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'version = "{version}"'
]
"README.md" = [
    'Latest version: {version}'
]
"src/tonydbc/__init__.py" = [
    '__version__ = "{version}"'
]
